Tout commentaire fonctionnel

+++ Domain/Commentary.cs
@@ -3,7 +3,7 @@
 public class Commentary
 {
     public int Id { get; set; }
-    public int note { get; set; }
+    public short note { get; set; }
     public string nom { get; set; }
     public string message { get; set; }
     public int id_User { get; set; }
